requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# SEC EDGAR
sec-edgar-downloader==5.0.0
lxml==4.9.4
beautifulsoup4==4.12.2

# Market Data
yfinance==0.2.33

# Machine Learning
scikit-learn==1.3.2
torch==2.1.2
torch-geometric==2.4.0

# Graph Analysis
networkx==3.2.1

# NLP
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Explainability
shap==0.44.0
lime==0.2.0.1

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Configuration & Logging
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
requests==2.31.0
